fix folder designed
