+++ public/locales/nl/common.json
@@ -80,7 +80,7 @@
     "sign-in": "Inloggen",
     "sign-up": "Inschrijven",
     "logout": "Uitloggen",
-    "my-profile": "Mijn profiel",
+    "my-profile": "Mijn Profiel",
     "user-name": "Gebruikersnaam",
     "email": "E-mail",
     "my-arts": "Mijn Kunsten"
